# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk mengelola data alumni dengan MongoDB menggunakan Clean Architecture",
        "title": "CRUD App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
    "/api/login": {
      "post": {
        "summary": "Login user",
        "description": "Autentikasi user dan mengembalikan token JWT",
        "tags": ["Auth"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Email dan Password",
            "schema": {"type": "object"}
          }
        ],
        "responses": {
          "200": {"description": "Login berhasil"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/api/profile": {
      "get": {
        "summary": "Mendapatkan profil user yang sedang login",
        "description": "Mengambil data profil user berdasarkan token JWT",
        "tags": ["Auth"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "responses": {
          "200": {"description": "Profile user"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/unair/alumni": {
      "get": {
        "summary": "Mendapatkan daftar alumni dengan pagination dan filter",
        "tags": ["Alumni"],
        "produces": ["application/json"],
        "parameters": [
          {"name": "page", "in": "query", "type": "integer", "default": 1},
          {"name": "limit", "in": "query", "type": "integer", "default": 10},
          {"name": "sortBy", "in": "query", "type": "string"},
          {"name": "order", "in": "query", "type": "string"},
          {"name": "search", "in": "query", "type": "string"}
        ],
        "responses": {"200": {"description": "Daftar alumni"}}
      },
      "post": {
        "summary": "Menambah data alumni baru",
        "tags": ["Alumni"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "body", "in": "body", "required": true, "schema": {"type": "object"}}
        ],
        "responses": {"201": {"description": "Alumni berhasil ditambahkan"}}
      }
    },
    "/unair/alumni/{id}": {
      "get": {
        "summary": "Mendapatkan data alumni berdasarkan ID",
        "tags": ["Alumni"],
        "produces": ["application/json"],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Data alumni"}}
      },
      "put": {
        "summary": "Mengupdate data alumni",
        "tags": ["Alumni"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"},
          {"name": "body", "in": "body", "required": true, "schema": {"type": "object"}}
        ],
        "responses": {"200": {"description": "Alumni berhasil diupdate"}}
      },
      "delete": {
        "summary": "Menghapus alumni (soft delete)",
        "tags": ["Alumni"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Alumni berhasil dihapus"}}
      },
      "patch": {
        "summary": "Merestore alumni yang dihapus",
        "tags": ["Alumni"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Alumni berhasil direstore"}}
      }
    },
    "/unair/alumni/without-pekerjaan": {
      "get": {
        "summary": "Mendapatkan alumni tanpa pekerjaan",
        "tags": ["Alumni"],
        "produces": ["application/json"],
        "responses": {"200": {"description": "Daftar alumni tanpa pekerjaan"}}
      }
    },
    "/unair/pekerjaan-alumni": {
      "get": {
        "summary": "Mendapatkan daftar pekerjaan alumni",
        "tags": ["Pekerjaan"],
        "produces": ["application/json"],
        "responses": {"200": {"description": "Daftar pekerjaan"}}
      },
      "post": {
        "summary": "Menambah data pekerjaan baru",
        "tags": ["Pekerjaan"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "body", "in": "body", "required": true, "schema": {"type": "object"}}
        ],
        "responses": {"201": {"description": "Pekerjaan berhasil ditambahkan"}}
      }
    },
    "/unair/pekerjaan-alumni/{id}": {
      "get": {
        "summary": "Mendapatkan data pekerjaan berdasarkan ID",
        "tags": ["Pekerjaan"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Data pekerjaan"}}
      },
      "put": {
        "summary": "Mengupdate data pekerjaan",
        "tags": ["Pekerjaan"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"},
          {"name": "body", "in": "body", "required": true, "schema": {"type": "object"}}
        ],
        "responses": {"200": {"description": "Pekerjaan berhasil diupdate"}}
      },
      "delete": {
        "summary": "Menghapus data pekerjaan (soft delete)",
        "tags": ["Pekerjaan"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Pekerjaan berhasil dihapus"}}
      },
      "patch": {
        "summary": "Merestore data pekerjaan",
        "tags": ["Pekerjaan"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Pekerjaan berhasil direstore"}}
      }
    },
    "/unair/pekerjaan-alumni/trash": {
      "get": {
        "summary": "Mendapatkan daftar pekerjaan alumni yang dihapus (trash)",
        "tags": ["Pekerjaan"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "responses": {"200": {"description": "Daftar pekerjaan di trash"}}
      }
    },
    "/unair/pekerjaan-alumni/alumni/{alumni_id}": {
      "get": {
        "summary": "Mendapatkan pekerjaan berdasarkan ID alumni",
        "tags": ["Pekerjaan"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "alumni_id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Daftar pekerjaan alumni"}}
      }
    },
    "/unair/pekerjaan-alumni/restore/{id}": {
      "put": {
        "summary": "Merestore pekerjaan dari trash",
        "tags": ["Pekerjaan"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Pekerjaan berhasil direstore"}}
      }
    },
    "/unair/pekerjaan-alumni/trash/delete/{id}": {
      "delete": {
        "summary": "Menghapus permanen pekerjaan dari trash",
        "tags": ["Pekerjaan"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Pekerjaan berhasil dihapus permanen"}}
      }
    },
    "/files": {
      "get": {
        "summary": "Mendapatkan semua file yang diupload",
        "tags": ["Upload"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "responses": {"200": {"description": "Daftar file"}}
      }
    },
    "/files/upload/foto": {
      "post": {
        "summary": "Upload foto alumni",
        "tags": ["Upload"],
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "file", "in": "formData", "required": true, "type": "file"}
        ],
        "responses": {"201": {"description": "Foto berhasil diupload"}}
      }
    },
    "/files/upload/sertifikat": {
      "post": {
        "summary": "Upload sertifikat alumni",
        "tags": ["Upload"],
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "file", "in": "formData", "required": true, "type": "file"}
        ],
        "responses": {"201": {"description": "Sertifikat berhasil diupload"}}
      }
    },
    "/files/{id}": {
      "get": {
        "summary": "Mendapatkan file berdasarkan ID",
        "tags": ["Upload"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "Data file"}}
      },
      "delete": {
        "summary": "Menghapus file berdasarkan ID",
        "tags": ["Upload"],
        "produces": ["application/json"],
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {"200": {"description": "File berhasil dihapus"}}
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Token dengan prefix 'Bearer '"
    }
  }
}